// src/pages/tools/csv/swap-csv-columns/types.ts
export type InitialValuesType = {
  fromPositionStatus: boolean;
  fromPosition: string | '';
  toPositionStatus: boolean;
  toPosition: string | '';
  fromHeader: string | '';
  toHeader: string | '';
  emptyValuesFilling: boolean;
  customFiller: string | '';
  deleteComment: boolean;
  commentCharacter: string | '';
  emptyLines: boolean;
};